requirements.worker.txt
farm-haystack==1.21.2
huggingface-hub==0.16.4
sentencepiece==0.1.99
sentence-transformers==2.2.2
torch==2.0.1
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
asyncio==3.4.3
websockets==11.0.3
transformers==4.32.1
